requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiohttp-swagger==1.0.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.3.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
dataclass-factory==2.16
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.2
packaging==24.2
propcache==0.3.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
python-telegram-bot==21.11.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
setuptools==72.1.0
shazamio==0.8.0
shazamio_core==1.1.2
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.43.0
yarl==1.18.3
youtube-search==2.1.2
yt-dlp==2025.4.30
